requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
yfinance>=0.2.36
schedule>=1.2.0
pytz>=2024.1
PyYAML>=6.0.1
markdown2>=2.4.13
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0
